MSL: Rewrite how to deal with vertex/resource remapping.
We will need this since the original interface relies on having mutable
pointers being stored. This is very different from how all other
backends do things, so we take a hard break here, remove all these
interfaces, and rely on a cleaner "set_override" and then add some
queries. This is far easier to wrap in a C API and simplifies the API a
lot.

Also removes the 3 different binding indices for an override.
This appears not very useful since there can only be one active set/binding
combination in a shader stage anyways.